requirements.txt
firebase_admin==4.3.0
Flask==1.1.2
google_api_python_client==1.12.3
protobuf==3.13.0
